Allow Ignoring Ownership
It's possible that a controller can adopt managed resources by attaching
an owner reference to it.  This owner reference has the effect of
keeping the resource alive, even though our intention is to delete it.
This adds in a new annotation that inhibits the consideration of owner
references.

Fixes https://github.com/argoproj/argo-cd/issues/11972